pyproject.toml
[project]
name = "lakebase"
version = "0.1.0"
description = "Databricks Lakebase Python Package"
requires-python = ">=3.11"

dependencies = [
    "databricks-sdk>=0.87.0",
    "pyyaml==6.0.2"
]

[project.scripts]
lakebase-deploy = "lakebase.cli:main"

[tool.setuptools]
packages = ["lakebase"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"